requirements.txt
PyYAML>=6.0
pygame>=2.5.0

